multi_beta_binomial main

This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.

multi_beta_binomial_glm(
  .data,
  formula = ~1,
  .sample,
  .cell_type,
  .count,
  check_outliers = FALSE,
  approximate_posterior_inference = T,
  variance_association = F,
  cores = detectCores(),
  seed = sample(1:99999, size = 1),
  verbose = FALSE
)

Arguments

.data
A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvaue column | a significance column

formula
A formula. The sample formula used to perform the differential cell_type abundance analysis

.sample
A column name as symbol. The sample identifier

.cell_type
A column name as symbol. The cell_type identifier

.count
A column name as symbol. The cell_type abundance (read count)

check_outliers
A boolean. Whether to check for outliers before the fit.

approximate_posterior_inference
A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.

variance_association
A boolean. Whether the variance should be associated to the factor of interest.

cores
An integer. How many cored to be used with parallel calculations.

seed
An integer. Used for development and testing purposes

verbose
A boolean. Prints progression.

Value

A nested tibble tbl with cell_type-wise information: sample wise data | plot | ppc samples failed | exposure deleterious outliers