multi_beta_glm main

This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.

multi_beta_glm(
  .data,
  formula = ~1,
  .sample,
  check_outliers = FALSE,
  approximate_posterior_inference = TRUE,
  cores = detect_cores(),
  seed = sample(1e+05, 1)
)

Arguments

.data

A tibble including a cell_type name column | sample name column | read counts column | factor columns | Pvaue column | a significance column

formula

A formula. The sample formula used to perform the differential cell_type abundance analysis

.sample

A column name as symbol. The sample identifier

check_outliers

A boolean. Whether to check for outliers before the fit.

approximate_posterior_inference

A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.

cores

An integer. How many cored to be used with parallel calculations.

seed

An integer. Used for development and testing purposes

Value

A nested tibble tbl with cell_type-wise information: sample wise data | plot | ppc samples failed | exposure deleterious outliers