remove_unwanted_variation

This function uses the model to remove unwanted variation from a dataset using the estimated of the model. For example if you fit your data with this formula ~ factor_1 + factor_2 and use this formula to remove unwanted variation ~ factor_1, the factor_2 will be factored out.

remove_unwanted_variation(
  .data,
  formula_composition = ~1,
  formula_variability = NULL
)

Arguments

.data

A tibble. The result of sccomp_glm.

formula_composition

A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

formula_variability

A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each factors. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

  estimates = sccomp_glm(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  )
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)

  remove_unwanted_variation(estimates)
#> sccomp says: calculating residuals
#> sccomp says: regressing out unwanted factors
#> # A tibble: 720 × 5
#>    sample       cell_group adjusted_proportion adjusted_counts logit_residuals
#>    <chr>        <chr>                    <dbl>           <dbl>           <dbl>
#>  1 10x_6K       B1                      0.0706           355.        0.245    
#>  2 10x_8K       B1                      0.0653           528.        0.163    
#>  3 GSE115189    B1                      0.0534           134.       -0.0518   
#>  4 SCP345_580   B1                      0.0427           263.       -0.287    
#>  5 SCP345_860   B1                      0.0538           370.       -0.0448   
#>  6 SCP424_pbmc1 B1                      0.0560           160.        0.0000171
#>  7 SCP424_pbmc2 B1                      0.0996           319.        0.623    
#>  8 SCP591       B1                      0.0282            16.9      -0.717    
#>  9 SI-GA-E5     B1                      0.0327           156.       -0.639    
#> 10 SI-GA-E7     B1                      0.123            943.        0.963    
#> # … with 710 more rows