replicate_data

This function replicates counts from a real-world dataset.

replicate_data(.data, number_of_draws = 1, mcmc_seed = sample(1e+05, 1))

Arguments

.data

A tibble including a cell_group name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column

number_of_draws

An integer. How may copies of the data you want to draw from the model joint posterior distribution.

mcmc_seed

An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

  sccomp_glm(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  ) |>

  replicate_data()
#> sccomp says: estimation [ETA: ~20s]
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> # A tibble: 720 × 5
#>    generated_proportions generated_counts replicate sample       cell_group
#>                    <dbl>            <int>     <int> <chr>        <chr>     
#>  1               0.0600               301         1 10x_6K       B1        
#>  2               0.0429               346         1 10x_8K       B1        
#>  3               0.0551               138         1 GSE115189    B1        
#>  4               0.0655               403         1 SCP345_580   B1        
#>  5               0.0720               495         1 SCP345_860   B1        
#>  6               0.0125                35         1 SCP424_pbmc1 B1        
#>  7               0.0766               245         1 SCP424_pbmc2 B1        
#>  8               0.0575                34         1 SCP591       B1        
#>  9               0.0543               259         1 SI-GA-E5     B1        
#> 10               0.00884               67         1 SI-GA-E7     B1        
#> # … with 710 more rows