sccomp_glm main

This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.

sccomp_glm(
  .data,
  formula = ~1,
  .sample,
  .cell_group,
  .count = NULL,
  prior_mean_variable_association = list(intercept = c(4.436925, 1.304049), slope =
    c(-0.73074903, 0.06532897), standard_deviation = c(0.4527292, 0.3318759)),
  percent_false_positive = 5,
  check_outliers = TRUE,
  approximate_posterior_inference = FALSE,
  test_composition_above_logit_fold_change = 0.2,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  variance_association = FALSE,
  cores = detectCores(),
  mcmc_seed = sample(1e+05, 1)
)

Arguments

.data

A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column

formula

A formula. The sample formula used to perform the differential cell_type abundance analysis

.sample

A column name as symbol. The sample identifier

.cell_group

A column name as symbol. The cell_type identifier

.count

A column name as symbol. The cell_type abundance (read count). Used only for data frame count output.

prior_mean_variable_association

A list of the form list(intercept = c(4.436925, 1.304049), slope = c(-0.73074903, 0.06532897), standard_deviation = c(0.4527292, 0.3318759)). Where for each parameter, we specify mean and standard deviation. This is used to incorporate prior knowledge about the mean/variability association of cell-type proportions.

percent_false_positive

A real between 0 and 100. It is the aimed percent of cell types being a false positive. For example, percent_false_positive_genes = 1 provide 1 percent of the calls for significant changes that are actually not significant.

check_outliers

A boolean. Whether to check for outliers before the fit.

approximate_posterior_inference

A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.

test_composition_above_logit_fold_change

A positive integer. It is the effect threshold used for the hypothesis test. A value of 0.2 correspond to a change in cell proportion of 10% for a cell type with baseline proportion of 50%. That is, a cell type goes from 45% to 50%. When the baseline proportion is closer to 0 or 1 this effect thrshold has consistent value in the logit uncontrained scale.

verbose

A boolean. Prints progression.

noise_model

A character string. The two noise models available are multi_beta_binomial (default) and dirichlet_multinomial.

variance_association

A boolean. Whether the variance should depend on the factor of interest.

cores

An integer. How many cored to be used with parallel calculations.

mcmc_seed

An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = TRUE,
    check_outliers = FALSE,
    cores = 1
  )
#> sccomp says: estimation [ETA: ~20s]