test_contrasts

This function test ocntrasts from a sccomp result.

test_contrasts(
  .data,
  contrasts,
  percent_false_positive = 5,
  test_composition_above_logit_fold_change = 0.2
)

Arguments

.data

A tibble. The result of sccomp_glm.

contrasts

A vector of character strings. For example if your formula is ~ 0 + treatment and the covariate treatment has values yes and no, your contrast could be "constrasts = c(treatmentyes - treatmentno)".

percent_false_positive

A real between 0 and 100 non included. This used to identify outliers with a specific false positive rate.

test_composition_above_logit_fold_change

A positive integer. It is the effect threshold used for the hypothesis test. A value of 0.2 correspond to a change in cell proportion of 10% for a cell type with baseline proportion of 50%. That is, a cell type goes from 45% to 50%. When the baseline proportion is closer to 0 or 1 this effect thrshold has consistent value in the logit uncontrained scale.

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

  estimates =
  sccomp_glm(
  counts_obj ,
   ~ 0 + type, ~1,  sample, cell_group, count,
    check_outliers = FALSE,
    cores = 1
  ) |>

  test_contrasts("typecancer - typebenign")
#> sccomp says: estimation [ETA: ~20s]
#> sccomp says: the composition design matrix has columns: typebenign, typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)