plot

This function plots a summary of the results of the model.

# S3 method for sccomp_tbl
plot(x, ...)

Arguments

x

A tibble including a cell_group name column | sample name column | read counts column | factor columns | Pvalue column | a significance column

...

parameters like significance_threshold A real. FDR threshold for labelling significant cell-groups.

Value

A ggplot

Examples

data("counts_obj")

estimate =
  sccomp_estimate(
  counts_obj ,
   ~ type, ~1, sample, cell_group, count,
    approximate_posterior_inference = "all",
    cores = 1
  )
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000446 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 4.46 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100        -4651.641             1.000            1.000
#> Chain 1:    200        -3771.589             0.617            1.000
#> Chain 1:    300        -2605.960             0.560            0.447
#> Chain 1:    400        -2136.113             0.475            0.447
#> Chain 1:    500        -2054.279             0.388            0.233
#> Chain 1:    600        -2101.796             0.327            0.233
#> Chain 1:    700        -2901.856             0.320            0.233
#> Chain 1:    800        -3697.934             0.307            0.233
#> Chain 1:    900        -3153.888             0.292            0.220
#> Chain 1:   1000        -2542.331             0.287            0.233
#> Chain 1:   1100        -2275.686             0.198            0.220
#> Chain 1:   1200        -3496.451             0.210            0.220
#> Chain 1:   1300        -2931.565             0.185            0.215
#> Chain 1:   1400        -3455.575             0.178            0.193
#> Chain 1:   1500        -3391.233             0.176            0.193
#> Chain 1:   1600        -2956.381             0.188            0.193
#> Chain 1:   1700        -2711.348             0.170            0.173
#> Chain 1:   1800        -2335.485             0.164            0.161
#> Chain 1:   1900        -2999.141             0.169            0.161
#> Chain 1:   2000        -1881.186             0.204            0.161
#> Chain 1:   2100        -2365.802             0.213            0.193
#> Chain 1:   2200        -1872.940             0.205            0.193
#> Chain 1:   2300        -1561.403             0.205            0.200
#> Chain 1:   2400        -1663.885             0.196            0.200
#> Chain 1:   2500        -1185.433             0.235            0.205
#> Chain 1:   2600        -2325.203             0.269            0.221
#> Chain 1:   2700        -2436.454             0.265            0.221
#> Chain 1:   2800        -1846.087             0.280            0.263
#> Chain 1:   2900        -1426.938             0.288            0.294
#> Chain 1:   3000        -2249.504             0.265            0.294
#> Chain 1:   3100         -900.242             0.394            0.320
#> Chain 1:   3200        -1091.899             0.385            0.320
#> Chain 1:   3300        -1704.263             0.401            0.359
#> Chain 1:   3400        -2836.741             0.435            0.366
#> Chain 1:   3500        -2979.240             0.400            0.359
#> Chain 1:   3600        -2460.002             0.372            0.320
#> Chain 1:   3700        -2008.220             0.390            0.320
#> Chain 1:   3800        -1249.968             0.418            0.359
#> Chain 1:   3900        -2255.901             0.433            0.366
#> Chain 1:   4000        -2295.965             0.399            0.359
#> Chain 1:   4100        -1331.575             0.321            0.359
#> Chain 1:   4200        -1809.374             0.330            0.359
#> Chain 1:   4300        -2481.796             0.321            0.271
#> Chain 1:   4400        -1526.789             0.344            0.271
#> Chain 1:   4500        -1650.022             0.347            0.271
#> Chain 1:   4600        -1973.221             0.342            0.271
#> Chain 1:   4700        -2784.546             0.348            0.291
#> Chain 1:   4800        -2010.706             0.326            0.291
#> Chain 1:   4900        -2058.153             0.284            0.271
#> Chain 1:   5000        -2938.123             0.312            0.291
#> Chain 1:   5100        -2520.204             0.256            0.271
#> Chain 1:   5200        -2509.064             0.230            0.271
#> Chain 1:   5300        -1587.966             0.261            0.291
#> Chain 1:   5400        -1823.528             0.212            0.166
#> Chain 1:   5500        -1534.314             0.223            0.188
#> Chain 1:   5600        -1480.198             0.210            0.188
#> Chain 1:   5700        -1333.897             0.192            0.166
#> Chain 1:   5800        -1236.360             0.162            0.129
#> Chain 1:   5900        -1586.061             0.181            0.166
#> Chain 1:   6000        -1579.802             0.152            0.129
#> Chain 1:   6100        -1833.765             0.149            0.129
#> Chain 1:   6200         -780.476             0.284            0.138
#> Chain 1:   6300        -1802.957             0.282            0.138
#> Chain 1:   6400        -1914.923             0.275            0.138
#> Chain 1:   6500        -1881.450             0.258            0.110
#> Chain 1:   6600        -1388.838             0.290            0.138
#> Chain 1:   6700        -2162.587             0.315            0.220
#> Chain 1:   6800        -2109.462             0.309            0.220
#> Chain 1:   6900        -1812.282             0.304            0.164
#> Chain 1:   7000        -1269.851             0.346            0.355
#> Chain 1:   7100         -884.249             0.376            0.358
#> Chain 1:   7200        -1570.335             0.285            0.358
#> Chain 1:   7300         -728.634             0.343            0.358
#> Chain 1:   7400        -2093.581             0.403            0.427
#> Chain 1:   7500        -2429.548             0.415            0.427
#> Chain 1:   7600        -2476.137             0.381            0.427
#> Chain 1:   7700        -1739.427             0.388            0.427
#> Chain 1:   7800        -2116.585             0.403            0.427
#> Chain 1:   7900        -2625.247             0.406            0.427
#> Chain 1:   8000        -1883.053             0.403            0.424
#> Chain 1:   8100        -2355.815             0.379            0.394
#> Chain 1:   8200         -988.347             0.474            0.394
#> Chain 1:   8300        -2096.943             0.411            0.394
#> Chain 1:   8400        -1464.071             0.389            0.394
#> Chain 1:   8500        -1716.113             0.390            0.394
#> Chain 1:   8600        -1553.385             0.399            0.394
#> Chain 1:   8700        -1888.759             0.374            0.201
#> Chain 1:   8800        -2528.690             0.382            0.253
#> Chain 1:   8900        -2068.478             0.384            0.253
#> Chain 1:   9000        -1275.865             0.407            0.253
#> Chain 1:   9100        -1181.388             0.395            0.253
#> Chain 1:   9200        -1738.683             0.289            0.253
#> Chain 1:   9300        -2301.569             0.260            0.245
#> Chain 1:   9400        -2732.455             0.233            0.222
#> Chain 1:   9500        -2066.123             0.250            0.245
#> Chain 1:   9600        -1535.463             0.275            0.253
#> Chain 1:   9700        -2207.188             0.287            0.304
#> Chain 1:   9800        -1355.679             0.325            0.321
#> Chain 1:   9900        -1601.630             0.318            0.321
#> Chain 1:   10000        -1811.181             0.267            0.304
#> Chain 1: Informational Message: The maximum number of iterations is reached! The algorithm may not have converged.
#> Chain 1: This variational approximation is not guaranteed to be meaningful.
#> Chain 1: 
#> Chain 1: Drawing a sample of size 4000 from the approximate posterior... 
#> Chain 1: COMPLETED.

# estimate |> plot()