plot

This function plots a summary of the results of the model.

# S3 method for sccomp_tbl
plot(x, ...)

Arguments

x

A tibble including a cell_group name column | sample name column | read counts column | factor columns | Pvalue column | a significance column

...

parameters like significance_threshold A real. FDR threshold for labelling significant cell-groups.

Value

A ggplot

Examples

data("counts_obj")

estimate =
  sccomp_estimate(
  counts_obj ,
   ~ type, ~1, sample, cell_group, count,
    cores = 1
  )
#> sccomp says: From version 1.7.7 the model by default is fit with the variational inference method (variational_inference = TRUE; much faster). For a full Bayesian inference (HMC method; the gold standard) use variational_inference = FALSE.
#> This message is displayed once per session.
#> sccomp says: count column is an integer. The sum-constrained beta binomial model will be used
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000471 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 4.71 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Iteration: 250 / 250 [100%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 0.1].
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100        -4869.430             1.000            1.000
#> Chain 1:    200        -4332.406             0.562            1.000
#> Chain 1:    300        -4122.950             0.392            0.124
#> Chain 1:    400        -3984.006             0.302            0.124
#> Chain 1:    500        -3916.971             0.245            0.051
#> Chain 1:    600        -3853.108             0.207            0.051
#> Chain 1:    700        -3810.597             0.179            0.035
#> Chain 1:    800        -3788.597             0.158            0.035
#> Chain 1:    900        -3761.092             0.141            0.017
#> Chain 1:   1000        -3746.001             0.127            0.017
#> Chain 1:   1100        -3729.718             0.028            0.017
#> Chain 1:   1200        -3721.476             0.015            0.011
#> Chain 1:   1300        -3711.701             0.011            0.007   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

# estimate |> plot()