Main Function for SCCOMP Estimate

The sccomp_estimate function performs linear modeling on a table of cell counts, which includes a cell-group identifier, sample identifier, integer count, and factors (continuous or discrete). The user can define a linear model with an input R formula, where the first factor is the factor of interest. Alternatively, sccomp accepts single-cell data containers (e.g., Seurat, SingleCellExperiment, cell metadata, or group-size) and derives the count data from cell metadata.

sccomp_estimate(
  .data,
  formula_composition = ~1,
  formula_variability = ~1,
  .sample,
  .cell_group,
  .count = NULL,
  cores = detectCores(),
  bimodal_mean_variability_association = FALSE,
  percent_false_positive = 5,
  variational_inference = TRUE,
  prior_mean = list(intercept = c(0, 1), coefficients = c(0, 1)),
  prior_overdispersion_mean_association = list(intercept = c(5, 2), slope = c(0, 0.6),
    standard_deviation = c(10, 20)),
  .sample_cell_group_pairs_to_exclude = NULL,
  verbose = TRUE,
  enable_loo = FALSE,
  noise_model = "multi_beta_binomial",
  exclude_priors = FALSE,
  use_data = TRUE,
  mcmc_seed = sample(1e+05, 1),
  max_sampling_iterations = 20000,
  pass_fit = TRUE,
  approximate_posterior_inference = NULL
)

Arguments

.data

A tibble including cell_group name column, sample name column, read counts column (optional depending on the input class), and factor columns.

formula_composition

A formula describing the model for differential abundance.

formula_variability

A formula describing the model for differential variability.

.sample

A column name as symbol for the sample identifier.

.cell_group

A column name as symbol for the cell_group identifier.

.count

A column name as symbol for the cell_group abundance (read count).

cores

Number of cores to use for parallel calculations.

bimodal_mean_variability_association

Boolean for modeling mean-variability as bimodal.

percent_false_positive

Real number between 0 and 100 for outlier identification.

variational_inference

Boolean for using variational Bayes for posterior inference. It is faster and convenient. Setting this argument to FALSE runs the full Bayesian (Hamiltonian Monte Carlo) inference, slower but it is the gold standard.

prior_mean

List with prior knowledge about mean distribution, they are the intercept and coefficient.

prior_overdispersion_mean_association

List with prior knowledge about mean/variability association.

.sample_cell_group_pairs_to_exclude

Column name with boolean for sample/cell-group pairs exclusion.

verbose

Boolean to print progression.

enable_loo

Boolean to enable model comparison using the LOO package.

noise_model

Character string for the noise model (e.g., 'multi_beta_binomial').

exclude_priors

Boolean to run a prior-free model.

use_data

Boolean to run the model data-free.

mcmc_seed

Integer for MCMC reproducibility.

max_sampling_iterations

Integer to limit maximum iterations for large datasets.

pass_fit

Boolean to include the Stan fit as attribute in the output.

approximate_posterior_inference

DEPRECATED please use the variational_inference argument.

Value

A nested tibble tbl, with the following columns

cell_group - column including the cell groups being tested

parameter - The parameter being estimated, from the design matrix dscribed with the input formula_composition and formula_variability

factor - The factor in the formula corresponding to the covariate, if exists (e.g. it does not exist in case og Intercept or contrasts, which usually are combination of parameters)

c_lower - lower (2.5%) quantile of the posterior distribution for a composition (c) parameter.

c_effect - mean of the posterior distribution for a composition (c) parameter.

c_upper - upper (97.5%) quantile of the posterior distribution fo a composition (c) parameter.

c_pH0 - Probability of the null hypothesis (no difference) for a composition (c). This is not a p-value.

c_FDR - False-discovery rate of the null hypothesis (no difference) for a composition (c).

c_n_eff - Effective sample size - the number of independent draws in the sample, the higher the better (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

c_R_k_hat - R statistic, a measure of chain equilibrium, should be within 0.05 of 1.0 (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

v_lower - Lower (2.5%) quantile of the posterior distribution for a variability (v) parameter

v_effect - Mean of the posterior distribution for a variability (v) parameter

v_upper - Upper (97.5%) quantile of the posterior distribution for a variability (v) parameter

v_pH0 - Probability of the null hypothesis (no difference) for a variability (v). This is not a p-value.

v_FDR - False-discovery rate of the null hypothesis (no difference), for a variability (v).

v_n_eff - Effective sample size for a variability (v) parameter - the number of independent draws in the sample, the higher the better (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

v_R_k_hat - R statistic for a variability (v) parameter, a measure of chain equilibrium, should be within 0.05 of 1.0 (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

count_data Nested input count data.

Examples

data("counts_obj")

estimate =
  sccomp_estimate(
  counts_obj ,
   ~ type,
   ~1,
   sample,
   cell_group,
   count,
    cores = 1
  )
#> sccomp says: count column is an integer. The sum-constrained beta binomial model will be used
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000314 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.14 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Iteration: 250 / 250 [100%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 0.1].
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100        -4873.930             1.000            1.000
#> Chain 1:    200        -4345.614             0.561            1.000
#> Chain 1:    300        -4151.689             0.389            0.122
#> Chain 1:    400        -4007.191             0.301            0.122
#> Chain 1:    500        -3937.252             0.244            0.047
#> Chain 1:    600        -3880.822             0.206            0.047
#> Chain 1:    700        -3833.754             0.178            0.036
#> Chain 1:    800        -3809.119             0.157            0.036
#> Chain 1:    900        -3779.277             0.140            0.018
#> Chain 1:   1000        -3760.398             0.127            0.018
#> Chain 1:   1100        -3743.666             0.027            0.015
#> Chain 1:   1200        -3727.673             0.016            0.012
#> Chain 1:   1300        -3717.491             0.011            0.008   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 4000 from the approximate posterior... 
#> Chain 1: COMPLETED.