sccomp_predict

This function replicates counts from a real-world dataset.

sccomp_predict(
  fit,
  formula_composition = NULL,
  new_data = NULL,
  number_of_draws = 500,
  mcmc_seed = sample(1e+05, 1)
)

Arguments

fit

The result of sccomp_estimate.

formula_composition

A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

new_data

A sample-wise data frame including the column that represent the factors in your formula. If you want to predict proportions for 10 samples, there should be 10 rows. T

number_of_draws

An integer. How may copies of the data you want to draw from the model joint posterior distribution.

mcmc_seed

An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

if(.Platform$OS.type == "unix")
  sccomp_estimate(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    cores = 1
  ) |>

  sccomp_predict()
#> sccomp says: count column is an integer. The sum-constrained beta binomial model will be used
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> 
#> SAMPLING FOR MODEL 'glm_multi_beta_binomial' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000319 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.19 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 4300 [  0%]  (Warmup)
#> Chain 1: Iteration:  301 / 4300 [  7%]  (Sampling)
#> Chain 1: Iteration: 1300 / 4300 [ 30%]  (Sampling)
#> Chain 1: Iteration: 2300 / 4300 [ 53%]  (Sampling)
#> Chain 1: Iteration: 3300 / 4300 [ 76%]  (Sampling)
#> Chain 1: Iteration: 4300 / 4300 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.691 seconds (Warm-up)
#> Chain 1:                19.539 seconds (Sampling)
#> Chain 1:                22.23 seconds (Total)
#> Chain 1: 
#> # A tibble: 720 × 5
#>    cell_group sample       proportion_mean proportion_lower proportion_upper
#>    <chr>      <fct>                  <dbl>            <dbl>            <dbl>
#>  1 B1         10x_6K                0.0580           0.0464           0.0716
#>  2 B1         10x_8K                0.0580           0.0464           0.0716
#>  3 B1         GSE115189             0.0580           0.0464           0.0716
#>  4 B1         SCP345_580            0.0580           0.0464           0.0716
#>  5 B1         SCP345_860            0.0580           0.0464           0.0716
#>  6 B1         SCP424_pbmc1          0.0580           0.0464           0.0716
#>  7 B1         SCP424_pbmc2          0.0580           0.0464           0.0716
#>  8 B1         SCP591                0.0580           0.0464           0.0716
#>  9 B1         SI-GA-E5              0.0292           0.0210           0.0392
#> 10 B1         SI-GA-E7              0.0292           0.0210           0.0392
#> # ℹ 710 more rows