sccomp_remove_outliers main

The function for linear modelling takes as input a table of cell counts with three columns containing a cell-group identifier, sample identifier, integer count and the factors (continuous or discrete). The user can define a linear model with an input R formula, where the first factor is the factor of interest. Alternatively, sccomp accepts single-cell data containers (Seurat, SingleCellExperiment44, cell metadata or group-size). In this case, sccomp derives the count data from cell metadata.

sccomp_remove_outliers(
  .estimate,
  percent_false_positive = 5,
  cores = detectCores(),
  variational_inference = FALSE,
  verbose = TRUE,
  mcmc_seed = sample(1e+05, 1),
  max_sampling_iterations = 20000,
  enable_loo = FALSE,
  approximate_posterior_inference = NULL
)

Arguments

.estimate

A tibble including a cell_group name column | sample name column | read counts column (optional depending on the input class) | factor columns.

percent_false_positive

A real between 0 and 100 non included. This used to identify outliers with a specific false positive rate.

cores

An integer. How many cored to be used with parallel calculations.

variational_inference

Boolean for using variational Bayes for posterior inference. It is faster and convenient. Setting this argument to FALSE runs the full Bayesian (Hamiltonian Monte Carlo) inference, slower but it is the gold standard.

verbose

A boolean. Prints progression.

mcmc_seed

An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()

max_sampling_iterations

An integer. This limit the maximum number of iterations in case a large dataset is used, for limiting the computation time.

enable_loo

A boolean. Enable model comparison by the R package LOO. This is helpful when you want to compare the fit between two models, for example, analogously to ANOVA, between a one factor model versus a interceot-only model.

approximate_posterior_inference

DEPRECATED please use the variational_inference argument.

Value

A nested tibble tbl, with the following columns

cell_group - column including the cell groups being tested

parameter - The parameter being estimated, from the design matrix dscribed with the input formula_composition and formula_variability

factor - The factor in the formula corresponding to the covariate, if exists (e.g. it does not exist in case og Intercept or contrasts, which usually are combination of parameters)

c_lower - lower (2.5%) quantile of the posterior distribution for a composition (c) parameter.

c_effect - mean of the posterior distribution for a composition (c) parameter.

c_upper - upper (97.5%) quantile of the posterior distribution fo a composition (c) parameter.

c_n_eff - Effective sample size - the number of independent draws in the sample, the higher the better (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

c_R_k_hat - R statistic, a measure of chain equilibrium, should be within 0.05 of 1.0 (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

v_lower - Lower (2.5%) quantile of the posterior distribution for a variability (v) parameter

v_effect - Mean of the posterior distribution for a variability (v) parameter

v_upper - Upper (97.5%) quantile of the posterior distribution for a variability (v) parameter

v_n_eff - Effective sample size for a variability (v) parameter - the number of independent draws in the sample, the higher the better (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

v_R_k_hat - R statistic for a variability (v) parameter, a measure of chain equilibrium, should be within 0.05 of 1.0 (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

count_data Nested input count data.

Examples

data("counts_obj")

estimate =
  sccomp_estimate(
  counts_obj ,
   ~ type,
   ~1,
   sample,
   cell_group,
   count,
    cores = 1
  ) |>
  sccomp_remove_outliers(cores = 1)
#> sccomp says: count column is an integer. The sum-constrained beta binomial model will be used
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> 
#> SAMPLING FOR MODEL 'glm_multi_beta_binomial' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000326 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.26 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 4300 [  0%]  (Warmup)
#> Chain 1: Iteration:  301 / 4300 [  7%]  (Sampling)
#> Chain 1: Iteration: 1300 / 4300 [ 30%]  (Sampling)
#> Chain 1: Iteration: 2300 / 4300 [ 53%]  (Sampling)
#> Chain 1: Iteration: 3300 / 4300 [ 76%]  (Sampling)
#> Chain 1: Iteration: 4300 / 4300 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.736 seconds (Warm-up)
#> Chain 1:                19.46 seconds (Sampling)
#> Chain 1:                22.196 seconds (Total)
#> Chain 1: 
#> sccomp says: to do hypothesis testing run `sccomp_test()`, the `test_composition_above_logit_fold_change` = 0.1 equates to a change of ~10%
#> sccomp says: outlier identification - step 1/2
#> 
#> SAMPLING FOR MODEL 'glm_multi_beta_binomial' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000326 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.26 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:     1 / 20299 [  0%]  (Warmup)
#> Chain 1: Iteration:   301 / 20299 [  1%]  (Sampling)
#> Chain 1: Iteration:  1300 / 20299 [  6%]  (Sampling)
#> Chain 1: Iteration:  2300 / 20299 [ 11%]  (Sampling)
#> Chain 1: Iteration:  3300 / 20299 [ 16%]  (Sampling)
#> Chain 1: Iteration:  4300 / 20299 [ 21%]  (Sampling)
#> Chain 1: Iteration:  5300 / 20299 [ 26%]  (Sampling)
#> Chain 1: Iteration:  6300 / 20299 [ 31%]  (Sampling)
#> Chain 1: Iteration:  7300 / 20299 [ 35%]  (Sampling)
#> Chain 1: Iteration:  8300 / 20299 [ 40%]  (Sampling)
#> Chain 1: Iteration:  9300 / 20299 [ 45%]  (Sampling)
#> Chain 1: Iteration: 10300 / 20299 [ 50%]  (Sampling)
#> Chain 1: Iteration: 11300 / 20299 [ 55%]  (Sampling)
#> Chain 1: Iteration: 12300 / 20299 [ 60%]  (Sampling)
#> Chain 1: Iteration: 13300 / 20299 [ 65%]  (Sampling)
#> Chain 1: Iteration: 14300 / 20299 [ 70%]  (Sampling)
#> Chain 1: Iteration: 15300 / 20299 [ 75%]  (Sampling)
#> Chain 1: Iteration: 16300 / 20299 [ 80%]  (Sampling)
#> Chain 1: Iteration: 17300 / 20299 [ 85%]  (Sampling)
#> Chain 1: Iteration: 18300 / 20299 [ 90%]  (Sampling)
#> Chain 1: Iteration: 19300 / 20299 [ 95%]  (Sampling)
#> Chain 1: Iteration: 20299 / 20299 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 3.042 seconds (Warm-up)
#> Chain 1:                96.186 seconds (Sampling)
#> Chain 1:                99.228 seconds (Total)
#> Chain 1: 
#> sccomp says: outlier-free model fitting - step 2/2
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> 
#> SAMPLING FOR MODEL 'glm_multi_beta_binomial' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000328 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.28 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 4300 [  0%]  (Warmup)
#> Chain 1: Iteration:  301 / 4300 [  7%]  (Sampling)
#> Chain 1: Iteration: 1300 / 4300 [ 30%]  (Sampling)
#> Chain 1: Iteration: 2300 / 4300 [ 53%]  (Sampling)
#> Chain 1: Iteration: 3300 / 4300 [ 76%]  (Sampling)
#> Chain 1: Iteration: 4300 / 4300 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 3.005 seconds (Warm-up)
#> Chain 1:                19.503 seconds (Sampling)
#> Chain 1:                22.508 seconds (Total)
#> Chain 1: