sccomp_remove_unwanted_variation

This function uses the model to remove unwanted variation from a dataset using the estimated of the model. For example if you fit your data with this formula ~ factor_1 + factor_2 and use this formula to remove unwanted variation ~ factor_1, the factor_2 will be factored out.

sccomp_remove_unwanted_variation(
  .data,
  formula_composition = ~1,
  formula_variability = NULL
)

Arguments

.data

A tibble. The result of sccomp_estimate.

formula_composition

A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

formula_variability

A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each factors. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

  estimates = sccomp_estimate(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    cores = 1
  )
#> sccomp says: count column is an integer. The sum-constrained beta binomial model will be used
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> 
#> SAMPLING FOR MODEL 'glm_multi_beta_binomial' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000318 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.18 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 4300 [  0%]  (Warmup)
#> Chain 1: Iteration:  301 / 4300 [  7%]  (Sampling)
#> Chain 1: Iteration: 1300 / 4300 [ 30%]  (Sampling)
#> Chain 1: Iteration: 2300 / 4300 [ 53%]  (Sampling)
#> Chain 1: Iteration: 3300 / 4300 [ 76%]  (Sampling)
#> Chain 1: Iteration: 4300 / 4300 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.814 seconds (Warm-up)
#> Chain 1:                19.687 seconds (Sampling)
#> Chain 1:                22.501 seconds (Total)
#> Chain 1: 

  sccomp_remove_unwanted_variation(estimates)
#> sccomp says: calculating residuals
#> sccomp says: regressing out unwanted factors
#> # A tibble: 720 × 5
#>    sample       cell_group adjusted_proportion adjusted_counts logit_residuals
#>    <chr>        <chr>                    <dbl>           <dbl>           <dbl>
#>  1 10x_6K       B1                      0.0709           357.           0.230 
#>  2 10x_8K       B1                      0.0658           531.           0.148 
#>  3 GSE115189    B1                      0.0537           135.          -0.0668
#>  4 SCP345_580   B1                      0.0430           265.          -0.302 
#>  5 SCP345_860   B1                      0.0541           372.          -0.0598
#>  6 SCP424_pbmc1 B1                      0.0564           162.          -0.0150
#>  7 SCP424_pbmc2 B1                      0.100            320.           0.608 
#>  8 SCP591       B1                      0.0283            17.0         -0.732 
#>  9 SI-GA-E5     B1                      0.0324           155.          -0.673 
#> 10 SI-GA-E7     B1                      0.121            925.           0.929 
#> # ℹ 710 more rows