sccomp_remove_unwanted_variation

This function uses the model to remove unwanted variation from a dataset using the estimated of the model. For example if you fit your data with this formula ~ factor_1 + factor_2 and use this formula to remove unwanted variation ~ factor_1, the factor_2 will be factored out.

sccomp_remove_unwanted_variation(
  .data,
  formula_composition = ~1,
  formula_variability = NULL
)

Arguments

.data

A tibble. The result of sccomp_estimate.

formula_composition

A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

formula_variability

A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each factors. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

  estimates = sccomp_estimate(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    cores = 1
  )
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00032 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.2 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100        -4050.430             1.000            1.000
#> Chain 1:    200        -3709.809             0.546            1.000
#> Chain 1:    300        -3704.246             0.364            0.092
#> Chain 1:    400        -3695.634             0.274            0.092
#> Chain 1:    500        -3698.656             0.219            0.002   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 4000 from the approximate posterior... 
#> Chain 1: COMPLETED.

  sccomp_remove_unwanted_variation(estimates)
#> sccomp says: calculating residuals
#> sccomp says: regressing out unwanted factors
#> # A tibble: 720 × 5
#>    sample       cell_group adjusted_proportion adjusted_counts logit_residuals
#>    <chr>        <chr>                    <dbl>           <dbl>           <dbl>
#>  1 10x_6K       B1                      0.0717           361.           0.230 
#>  2 10x_8K       B1                      0.0664           537.           0.148 
#>  3 GSE115189    B1                      0.0543           136.          -0.0673
#>  4 SCP345_580   B1                      0.0434           268.          -0.302 
#>  5 SCP345_860   B1                      0.0547           376.          -0.0602
#>  6 SCP424_pbmc1 B1                      0.0570           163.          -0.0154
#>  7 SCP424_pbmc2 B1                      0.101            323.           0.607 
#>  8 SCP591       B1                      0.0287            17.2         -0.732 
#>  9 SI-GA-E5     B1                      0.0337           161.          -0.622 
#> 10 SI-GA-E7     B1                      0.125            960.           0.979 
#> # ℹ 710 more rows