sccomp_replicate

This function replicates counts from a real-world dataset.

sccomp_replicate(
  fit,
  formula_composition = NULL,
  formula_variability = NULL,
  number_of_draws = 1,
  mcmc_seed = sample(1e+05, 1)
)

Arguments

fit

The result of sccomp_estimate.

formula_composition

A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

formula_variability

A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each factors. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.

number_of_draws

An integer. How may copies of the data you want to draw from the model joint posterior distribution.

mcmc_seed

An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

if(.Platform$OS.type == "unix")
  sccomp_estimate(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    cores = 1
  ) |>

  sccomp_replicate()
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000308 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.08 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100        -3897.812             1.000            1.000
#> Chain 1:    200        -3713.862             0.525            1.000
#> Chain 1:    300        -3698.354             0.351            0.050
#> Chain 1:    400        -3699.842             0.264            0.050
#> Chain 1:    500        -3696.201             0.211            0.004   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 4000 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> # A tibble: 720 × 5
#>    cell_group sample       generated_proportions generated_counts replicate
#>    <chr>      <chr>                        <dbl>            <int>     <int>
#>  1 B1         10x_6K                      0.0513              256         1
#>  2 B1         10x_8K                      0.106               530         1
#>  3 B1         GSE115189                   0.0488              243         1
#>  4 B1         SCP345_580                  0.0493              246         1
#>  5 B1         SCP345_860                  0.0404              201         1
#>  6 B1         SCP424_pbmc1                0.0571              285         1
#>  7 B1         SCP424_pbmc2                0.0548              274         1
#>  8 B1         SCP591                      0.0795              397         1
#>  9 B1         SI-GA-E5                    0.0193               96         1
#> 10 B1         SI-GA-E7                    0.0587              293         1
#> # ℹ 710 more rows