simulate_data

This function simulates counts from a linear model.

simulate_data(
  .data,
  formula = ~1,
  .sample,
  .cell_group,
  .sample_cell_count,
  .coefficients,
  mean_variable_association = c(5.6260004, -0.6940178, 0.816423129),
  percent_false_positive = 5,
  check_outliers = TRUE,
  approximate_posterior_inference = FALSE,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  cores = detectCores(),
  seed = 42
)

Arguments

.data
A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column

formula
A formula. The sample formula used to perform the differential cell_type abundance analysis

.sample
A column name as symbol. The sample identifier

.cell_group
A column name as symbol. The cell_type identifier

.sample_cell_count
A integer vector. The total number of cells for each sample.

.coefficients
A matrix of coefficients.

mean_variable_association
A numeric vector of size 3. The intercept, slope and standard deviation of the proportion mean/variability association.

percent_false_positive
A real between 0 and 100. It is the aimed percent of cell types being a false positive. For example, percent_false_positive_genes = 1 provide 1 percent of the calls for significant changes that are actually not significant.

check_outliers
A boolean. Whether to check for outliers before the fit.

approximate_posterior_inference
A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.

verbose
A boolean. Prints progression.

noise_model
A character string. The two noise models available are multi_beta_binomial (default) and dirichlet_multinomial.

cores
An integer. How many cored to be used with parallel calculations.

seed
An integer. Used for development and testing purposes

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = FALSE
  )
#> sccomp says: outlier identification first pass - step 1/3 [ETA: ~20s]
#> sccomp says: outlier identification second pass - step 2/3 [ETA: ~60s]
#> sccomp says: outlier-free model fitting - step 3/3 [ETA: ~20s]