simulate_data

This function simulates counts from a linear model.

simulate_data(
  .data,
  .estimate_object,
  formula_composition,
  formula_variability = NULL,
  .sample = NULL,
  .cell_group = NULL,
  .coefficients = NULL,
  variability_multiplier = 5,
  number_of_draws = 1,
  mcmc_seed = sample(1e+05, 1)
)

Arguments

.data

A tibble including a cell_group name column | sample name column | read counts column | factor columns | Pvalue column | a significance column

.estimate_object

The result of sccomp_estimate execution. This is used for sampling from real-data properties.

formula_composition

A formula. The sample formula used to perform the differential cell_group abundance analysis

formula_variability

A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each factors.

.sample

A column name as symbol. The sample identifier

.cell_group

A column name as symbol. The cell_group identifier

.coefficients

The column names for coefficients, for example, c(b_0, b_1)

variability_multiplier

A real scalar. This can be used for artificially increasing the variability of the simulation for benchmarking purposes.

number_of_draws

An integer. How may copies of the data you want to draw from the model joint posterior distribution.

mcmc_seed

An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()

Value

A nested tibble tbl with cell_group-wise statistics

Examples

data("counts_obj")
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

estimate =
 sccomp_estimate(
 counts_obj ,
  ~ type, ~1,  sample, cell_group, count,
   cores = 1
 )
#> sccomp says: count column is an integer. The sum-constrained beta binomial model will be used
#> sccomp says: estimation
#> sccomp says: the composition design matrix has columns: (Intercept), typecancer
#> sccomp says: the variability design matrix has columns: (Intercept)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000329 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.29 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1:    100        -5835.207             1.000            1.000
#> Chain 1:    200        -3818.818             0.764            1.000
#> Chain 1:    300        -3700.740             0.520            0.528
#> Chain 1:    400        -3696.675             0.390            0.528
#> Chain 1:    500        -3694.667             0.312            0.032
#> Chain 1:    600        -3697.763             0.260            0.032
#> Chain 1:    700        -3692.091             0.223            0.002   MEDIAN ELBO CONVERGED
#> Chain 1: 
#> Chain 1: Drawing a sample of size 1000 from the approximate posterior... 
#> Chain 1: COMPLETED.

# Set coefficients for cell_groups. In this case all coefficients are 0 for simplicity.
counts_obj = counts_obj |> mutate(b_0 = 0, b_1 = 0)
# Simulate data
simulate_data(counts_obj, estimate, ~type, ~1, sample, cell_group, c(b_0, b_1))
#> # A tibble: 720 × 10
#>    sample type   phenotype    count cell_group   b_0   b_1 generated_proportions
#>    <chr>  <fct>  <fct>        <int> <chr>      <dbl> <dbl>                 <dbl>
#>  1 10x_6K benign b_cell_macr…    42 BM             0     0               0.101  
#>  2 10x_6K benign B_cell:imma…   361 B1             0     0               0      
#>  3 10x_6K benign B_cell:imma…    57 B2             0     0               0      
#>  4 10x_6K benign B_cell:Memo…    40 B3             0     0               0      
#>  5 10x_6K benign Dendritic_C…    75 Dm             0     0               0      
#>  6 10x_6K benign HSC_-G-CSF_…   187 H1             0     0               0.429  
#>  7 10x_6K benign HSC_-G-CSF_…   269 H2             0     0               0      
#>  8 10x_6K benign HSC_Bcell_m…   190 H3             0     0               0      
#>  9 10x_6K benign Mixed_CD3_c…   123 TM1            0     0               0.00386
#> 10 10x_6K benign Mixed_CD3_c…    19 TM2            0     0               0      
#> # ℹ 710 more rows
#> # ℹ 2 more variables: generated_counts <int>, replicate <int>